+++ pyhilo/oauth2.py
@@ -1,4 +1,5 @@
 """Custom OAuth2 implementation."""
+
 import base64
 import hashlib
 import os
@@ -8,14 +9,9 @@ from typing import Any, cast
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.config_entry_oauth2_flow import LocalOAuth2Implementation
 
-from pyhilo.const import (
-    AUTH_AUTHORIZE,
-    AUTH_CHALLENGE_METHOD,
-    AUTH_CLIENT_ID,
-    AUTH_SCOPE,
-    AUTH_TOKEN,
-    DOMAIN,
-)
+from pyhilo.const import AUTH_AUTHORIZE, AUTH_CLIENT_ID, AUTH_TOKEN, DOMAIN
+
+from pyhilo.oauth2helper import OAuth2Helper
 
 
 class AuthCodeWithPKCEImplementation(LocalOAuth2Implementation):  # type: ignore[misc]
@@ -34,8 +30,8 @@ class AuthCodeWithPKCEImplementation(LocalOAuth2Implementation):  # type: ignore
             AUTH_AUTHORIZE,
             AUTH_TOKEN,
         )
-        self._code_verifier = self._get_code_verifier()
-        self._code_challenge = self._get_code_challange(self._code_verifier)
+
+        self.oauth_helper = OAuth2Helper()
 
     # ... Override AbstractOAuth2Implementation details
     @property
@@ -46,32 +42,15 @@ class AuthCodeWithPKCEImplementation(LocalOAuth2Implementation):  # type: ignore
     @property
     def extra_authorize_data(self) -> dict:
         """Extra data that needs to be appended to the authorize url."""
-        return {
-            "scope": AUTH_SCOPE,
-            "code_challenge": self._code_challenge,
-            "code_challenge_method": AUTH_CHALLENGE_METHOD,
-        }
+        return self.oauth_helper.get_authorize_parameters()
 
     async def async_resolve_external_data(self, external_data: Any) -> dict:
         """Resolve the authorization code to tokens."""
         return cast(
             dict,
             await self._token_request(
-                {
-                    "grant_type": "authorization_code",
-                    "code": external_data["code"],
-                    "redirect_uri": external_data["state"]["redirect_uri"],
-                    "code_verifier": self._code_verifier,
-                },
+                self.oauth_helper.get_token_request_parameters(
+                    external_data["code"], external_data["state"]["redirect_uri"]
+                )
             ),
         )
-
-    # Ref : https://blog.sanghviharshit.com/reverse-engineering-private-api-oauth-code-flow-with-pkce/
-    def _get_code_verifier(self) -> str:
-        code = base64.urlsafe_b64encode(os.urandom(40)).decode("utf-8")
-        return re.sub("[^a-zA-Z0-9]+", "", code)
-
-    def _get_code_challange(self, verifier: str) -> str:
-        sha_verifier = hashlib.sha256(verifier.encode("utf-8")).digest()
-        code = base64.urlsafe_b64encode(sha_verifier).decode("utf-8")
-        return code.replace("=", "")
